Attempt to implement an auto-eject system for reactors
Some naming changes would need to be done before pushing this to the mainline
(i.e. renaming active pushing to auto-eject since that is what mekanism uses
- no need to reinvent the wheel.). To be honest the presentation could be
improved quite a bit (for example have an extra tooltip that hints at the
fact that most piping mods have a wrench/configuration item that configures
inputs and outputs - me being an absolute dunce forgot that fact)

Also this system has troubles with mekanism transporters. They were the main
motivation of me drafting up this PR so I'm not too sure whether I'd like
to have this system be pushed to production as-is. On the other hand chests
and most other contains should work.

Automatically requesting input will probably not be implemented due
to performance concerns (mekanism doesn't do it so we probably shouldn't
either)

Performance-wise this system could be improved (i.e. don't attempt to push
itemstacks if there are no itemstacks to be pushed), but is most likely
"good enough".